requirements.txt
Flask==3.1.0
Flask-CORS==5.0.1
Pillow==11.1.0
pytesseract==0.3.13
gunicorn==23.0.0
python-dateutil==2.9.0
supabase==2.13.0
python-dotenv==1.1.0
requests==2.32.3
